fix: 型定義, baseUrl
