+++ next-app/Api/ApiConfig.ts
@@ -1,0 +1,2 @@
+
+export  const baseUrl = 'http://localhost:3001/'
